// completedProject/src/models/vehicle.model.ts
export class Vehicle {
    name:                   string;
    model:                  string;
    vehicle_class:          string;
    manufacturer:           string;
    length:                 string;
    cost_in_credits:        string;
    crew:                   string;
    passengers:             string;
    max_atmosphering_speed: string;
    cargo_capacity:         string;
    consumables:            string;
    created:                string;
    edited:                 string;
}